+++ db/DBManager.py
@@ -1,0 +1,14 @@
+import json
+
+class db_manager:
+    @staticmethod
+    def read(file_path):
+        with open(file_path, 'r') as f:
+            return json.load(f)
+
+    @staticmethod
+    def save(filename,data):
+        with open(filename, 'w') as f:
+            json.dump(data, f, indent=4)
+
+
