+++ db/data/variants.json
@@ -1,0 +1,14 @@
+[
+    {
+        "id": 0,
+        "name": "Color"
+    },
+    {
+        "id": 1,
+        "name": "Size"
+    },
+    {
+        "id": 2,
+        "name": "Length"
+    }
+]
